Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4707



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@20028 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/utext.cpp
@@ -1102,7 +1102,7 @@ swapBuffers:
         U_ASSERT(ix<=u8b->bufNativeLimit);
         mapIndex = ix - u8b->toUCharsMapStart;
         U_ASSERT(mapIndex>=0);
-        U_ASSERT(mapIndex<sizeof(u8b->mapToUChars));
+        U_ASSERT(mapIndex<(int32_t)sizeof(u8b->mapToUChars));
         ut->chunkOffset = u8b->mapToUChars[mapIndex] - u8b->bufStartIdx;
 
         return TRUE;
